Move `MainFormParamViewModel` to the viewModel folder.
